Fix `BoxedString` not having the right provenance
When creating a `BoxString` using `from_string` or `from_box_str`,
`str::as_ptr` was used to get the pointer, which only has read
provenance for the initialized part of the string. Going through `Vec`
in the `from_string` and through `Box::into_raw` in the `from_box_str`
case fixed the issue.