Rewrite `ArcString` to correctly handle provenance
Before, it used a zero length array to signal it being a fake DST.
This is not allowed, since it doesn't allow writing past the end
of this header. Rewrite the entire module to use an `ArcStringPtr`
abstraction instead, which allows to easily write to the buffer
and still have the header accessible.