Fix provenance issues (#100)
* Fix `BoxedString` not having the right provenance

When creating a `BoxString` using `from_string` or `from_box_str`,
`str::as_ptr` was used to get the pointer, which only has read
provenance for the initialized part of the string. Going through `Vec`
in the `from_string` and through `Box::into_raw` in the `from_box_str`
case fixed the issue.

* Rewrite `ArcString` to correctly handle provenance

Before, it used a zero length array to signal it being a fake DST.
This is not allowed, since it doesn't allow writing past the end
of this header. Rewrite the entire module to use an `ArcStringPtr`
abstraction instead, which allows to easily write to the buffer
and still have the header accessible.

* Correctly set `MIRIFLAGS` in CI to `-Zmiri-strict-provenance`

Before, the miriflags didn't work correctly because they had a typo.
Fix the typo, and also use the new strict-provenance option which
implies the two used before.

* Remove `repr::arc`

The module is unused and can therefore be removed.

Co-authored-by: Parker Timmerman <parker@parkertimmerman.com>